dedupe getPathWithExtension
