+++ src/parser/parse.js
@@ -9,8 +9,9 @@ import type {
 import {parse} from '@babel/parser'; // eslint-disable-line flowtype-errors/uncovered
 import traverse from '@babel/traverse'; // eslint-disable-line flowtype-errors/uncovered
 
-import fs from 'fs';
 import path from 'path';
+
+import {getPathWithExtension} from './utils';
 
 /**
  * This file is responsible for finding all gql`-annotated
@@ -75,29 +76,6 @@ export type FileResult = {|
 |};
 
 export type Files = {[path: string]: FileResult};
-
-export const getPathWithExtension = (pathWithoutExtension: string): string => {
-    if (
-        /\.(less|css|png|gif|jpg|jpeg|js|jsx|ts|tsx|mjs)$/.test(
-            pathWithoutExtension,
-        )
-    ) {
-        return pathWithoutExtension;
-    }
-    if (fs.existsSync(pathWithoutExtension + '.js')) {
-        return pathWithoutExtension + '.js';
-    }
-    if (fs.existsSync(pathWithoutExtension + '.jsx')) {
-        return pathWithoutExtension + '.jsx';
-    }
-    if (fs.existsSync(pathWithoutExtension + '.tsx')) {
-        return pathWithoutExtension + '.tsx';
-    }
-    if (fs.existsSync(pathWithoutExtension + '.ts')) {
-        return pathWithoutExtension + '.ts';
-    }
-    throw new Error("Can't find file at " + pathWithoutExtension);
-};
 
 /**
  * Finds all referenced imports that might possibly be relevant
